images repaired
